refactor: move funcoes de finalizacao pra store
