+++ src/pages/recover/request/index.tsx
@@ -1,41 +1,12 @@
 import TemplatePage from '@pages/templatePage';
 import { InputText } from 'primereact/inputtext';
-
-import './style.css';
 import { Button } from 'primereact/button';
 import FormField from '@components/FormField/formField';
 import { useRecoverRequest } from '@stores/recover/recoverRequest';
-import { useErrorContext } from '@contexts/errorContext';
-import { recuperar_senha } from 'routes/routesRecover';
-import { useNotification } from '@contexts/notificationContext';
+import './style.css';
 
 const RecoverRequestPage = () => {
-  const { credenciais, setField, validate } = useRecoverRequest();
-  const { setError } = useErrorContext();
-  const { showNotification } = useNotification();
-
-  const finalizeRecoverRequest = async () => {
-    if (validate()) {
-      try {
-        const response = await recuperar_senha(credenciais);
-
-        if (response.status === 200) {
-          showNotification('success', response.data.mensagem, '');
-        }
-      } catch (error) {
-        if (error.response) {
-          const errorMessage = error.response.data.message || 'Erro no servidor.';
-          showNotification('error', errorMessage, '');
-        } else if (error.request) {
-          showNotification('error', 'Sem resposta do servidor. Verifique sua conexão.', '');
-        } else {
-          showNotification('error', 'Algo deu errado. Tente novamente mais tarde.', '');
-        }
-      }
-    } else {
-      setError('email', { error: true, message: 'Por favor, preencha o campo.' });
-    }
-  };
+  const { credenciais, setField, finalizeRecoverRequest } = useRecoverRequest();
 
   return (
     <div className="main-recover-one">
@@ -63,7 +34,7 @@ const RecoverRequestPage = () => {
               className="button"
               type="submit"
               label="Enviar e-mail de recuperação"
-              onClick={finalizeRecoverRequest}
+              onClick={() => finalizeRecoverRequest()}
             />
             <div className="footer">
               <p>
