+++ src/stores/recover/recoverRequest.tsx
@@ -1,11 +1,15 @@
 import { createContext, ReactNode, useContext } from 'react';
 import { CredenciaisRecoverRequest } from '@domains/Credenciais';
 import { useForm } from '@hooks/useForm';
+import { recuperar_senha } from '@routes/routesRecover';
+import { useNotification } from '@contexts/notificationContext';
+import { useErrorContext } from '@contexts/errorContext';
 
 interface RecoverRequestContextType {
   credenciais: CredenciaisRecoverRequest;
   setField: (field: string, value: any) => void;
   validate: () => boolean;
+  finalizeRecoverRequest: () => void;
 }
 
 const RecoverRequestContext = createContext<RecoverRequestContextType | null>(null);
@@ -23,6 +27,8 @@ interface RecoverRequestProviderProps {
 }
 
 export const RecoverRequestProvider = ({ children }: RecoverRequestProviderProps) => {
+  const { showNotification } = useNotification();
+  const { setError } = useErrorContext();
   const { formData, setField, validate } = useForm<CredenciaisRecoverRequest>({
     initialData: {
       email: '',
@@ -32,8 +38,31 @@ export const RecoverRequestProvider = ({ children }: RecoverRequestProviderProps
     },
   });
 
+  const finalizeRecoverRequest = async () => {
+    if (validate()) {
+      try {
+        const response = await recuperar_senha(formData);
+
+        if (response.status === 200) {
+          showNotification('success', response.data.mensagem, '');
+        }
+      } catch (error: any) {
+        if (error.response) {
+          const errorMessage = error.response.data.message || 'Erro no servidor.';
+          showNotification('error', errorMessage, '');
+        } else if (error.request) {
+          showNotification('error', 'Sem resposta do servidor. Verifique sua conexão.', '');
+        } else {
+          showNotification('error', 'Algo deu errado. Tente novamente mais tarde.', '');
+        }
+      }
+    } else {
+      setError('email', { error: true, message: 'Por favor, preencha o campo.' });
+    }
+  };
+
   return (
-    <RecoverRequestContext.Provider value={{ credenciais: formData, setField, validate }}>
+    <RecoverRequestContext.Provider value={{ credenciais: formData, setField, validate, finalizeRecoverRequest }}>
       {children}
     </RecoverRequestContext.Provider>
   );
